MeroBazaar_ML/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.1
motor>=3.3.2
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
cmdstanpy>=1.2.0
